more scripts

+++ Assets/Player.cs
@@ -3,8 +3,8 @@ using System.Collections;
 
 public static class Player {
 	public static int laserCount = 10;
-	public static int gemCount = 100;
-	public static int moneyCount = 10;
+	public static float gemCount = 100;
+	public static float moneyCount = 10;
 	public static int goldRushCount = 2;
 	public static string name = "Meme MASTER";
